ProtocolSelector

+++ src/main/java/simple/net/protocol/ProtocolFactorySelector.java
@@ -2,5 +2,5 @@ package simple.net.protocol;
 
 public interface ProtocolFactorySelector {
 
-    ProtocolFactory select(int protocolCode);
+    ProtocolFactory select(byte protocolCode);
 }
